+++ crates/web-client/js/types/index.d.ts
@@ -51,7 +51,8 @@ export declare class WasmWebClient extends WasmWebClientBase {
     noteTransportUrl?: string,
     seed?: Uint8Array,
     storeName?: string,
-    logLevel?: LogLevel
+    logLevel?: LogLevel,
+    useWorker?: boolean
   ): Promise<WasmWebClient>;
 
   static createClientWithExternalKeystore(
@@ -62,7 +63,8 @@ export declare class WasmWebClient extends WasmWebClientBase {
     getKeyCb?: GetKeyCallback,
     insertKeyCb?: InsertKeyCallback,
     signCb?: SignCallback,
-    logLevel?: LogLevel
+    logLevel?: LogLevel,
+    useWorker?: boolean
   ): Promise<WasmWebClient>;
 
   syncState(): Promise<SyncSummary>;